# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "《一次买够》网购比价平台 API",
    "description": "一次买够网购比价平台的完整API文档，用于前后端接口对接。本文档定义了所有后端接口的请求/响应格式、数据结构与业务逻辑。",
    "version": "1.0.0",
    "contact": {
      "name": "后端开发团队",
      "email": "api-support@ycg.com",
      "url": "https://docs.ycg.com"
    },
    "license": {
      "name": "CC BY 4.0"
    },
    "x-api-id": "ycg-api",
    "x-api-lifecycle": "production"
  },
  "servers": [
    {
      "url": "https://api.ycg.com/api/v1",
      "description": "生产环境"
    },
    {
      "url": "https://test-api.ycg.com/api/v1",
      "description": "测试环境"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用Bearer token进行认证"
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "title": "通用API响应",
        "description": "所有API端点的统一响应格式",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码 (0=成功, 非0=失败)。详见错误码表",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "消息描述，成功时通常为'ok'，失败时为错误说明",
            "example": "ok",
            "maxLength": 256
          },
          "data": {
            "type": "object",
            "description": "响应数据正文，结构因不同接口而异。失败时通常为null",
            "nullable": true
          }
        },
        "required": ["code", "message"]
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            },
            "example": {
              "code": 400001,
              "message": "参数缺失",
              "data": null
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "微信登录",
        "description": "通过微信授权code获取access_token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "微信授权code"
                  },
                  "raw_data": {
                    "type": "string",
                    "description": "微信原始数据"
                  },
                  "signature": {
                    "type": "string",
                    "description": "微信签名"
                  }
                },
                "required": ["code", "raw_data", "signature"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "code": 0,
                  "message": "登录成功",
                  "data": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refresh_token": "refresh_token_value",
                    "expires_in": 86400,
                    "user_id": 12345,
                    "openid": "oZxsulsr5z0GQyVQW..."
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["认证"],
        "summary": "刷新Token",
        "description": "使用refresh_token获取新的access_token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "description": "刷新令牌"
                  }
                },
                "required": ["refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "code": 0,
                  "message": "刷新成功",
                  "data": {
                    "access_token": "new_access_token",
                    "refresh_token": "new_refresh_token",
                    "expires_in": 86400
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["认证"],
        "summary": "登出",
        "description": "用户登出",
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "code": 0,
                  "message": "登出成功",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/products/search": {
      "get": {
        "tags": ["商品"],
        "summary": "商品搜索",
        "description": "按关键词搜索商品，支持多维度筛选和排序",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "description": "搜索关键词，支持商品名称、品牌、型号等多维度搜索",
            "example": "iPhone 15 Pro Max"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "页码，从1开始",
            "example": 1
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "每页数量，最大100条",
            "example": 20
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "商品分类ID（可选），用于按类目筛选。如10001=电子产品，10002=服装等",
            "example": 10001
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "minimum": 0
            },
            "description": "最低价格（单位：元），用于价格区间筛选",
            "example": 5000.00
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "minimum": 0
            },
            "description": "最高价格（单位：元），用于价格区间筛选",
            "example": 9999.00
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "平台筛选，多个平台用英文逗号分隔。可选值：taobao,jingdong,pinduoduo,amazon,other",
            "example": "taobao,jingdong"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["price_asc", "price_desc", "sales_desc", "rating_desc"],
              "default": "price_asc"
            },
            "description": "排序方式：price_asc=低价优先，price_desc=高价优先，sales_desc=热销优先，rating_desc=高评优先",
            "example": "price_asc"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": ["商品"],
        "summary": "获取商品详情",
        "description": "获取单个商品的详细信息，包括多平台价格、历史价格等",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32
            },
            "description": "商品ID，由系统生成的唯一标识符，格式如SKU20250315001",
            "example": "SKU20250315001"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/products/categories": {
      "get": {
        "tags": ["商品"],
        "summary": "获取商品分类",
        "description": "获取商品分类树",
        "parameters": [
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            },
            "description": "父分类ID（0=获取顶级分类，其他值=获取该分类的子分类）",
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "tags": ["用户"],
        "summary": "获取用户信息",
        "description": "获取当前登录用户的信息",
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["用户"],
        "summary": "更新用户信息",
        "description": "更新当前用户的个人信息",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "更新用户信息请求",
                "properties": {
                  "nick_name": {
                    "type": "string",
                    "description": "昵称",
                    "example": "小明",
                    "maxLength": 64,
                    "minLength": 1,
                    "nullable": true
                  },
                  "avatar_url": {
                    "type": "string",
                    "format": "uri",
                    "description": "头像URL",
                    "example": "https://cdn.example.com/avatars/user123.jpg",
                    "maxLength": 512,
                    "nullable": true
                  },
                  "gender": {
                    "type": "integer",
                    "description": "性别（0=未知，1=男，2=女）",
                    "enum": [0, 1, 2],
                    "example": 1,
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/favorites": {
      "get": {
        "tags": ["收藏"],
        "summary": "获取收藏列表",
        "description": "获取当前用户的收藏列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "页码，从1开始",
            "example": 1
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "每页数量，最大100条",
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["收藏"],
        "summary": "添加收藏",
        "description": "将商品添加到用户收藏",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "添加收藏请求",
                "properties": {
                  "product_id": {
                    "type": "string",
                    "description": "商品ID",
                    "example": "SKU20250315001",
                    "maxLength": 32
                  },
                  "notes": {
                    "type": "string",
                    "description": "用户添加的收藏备注（可选），如'618打折再买'",
                    "example": "618打折再买",
                    "maxLength": 500,
                    "nullable": true
                  }
                },
                "required": ["product_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "收藏成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/favorites/{favorite_id}": {
      "delete": {
        "tags": ["收藏"],
        "summary": "移除收藏",
        "description": "取消商品收藏",
        "parameters": [
          {
            "name": "favorite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消收藏成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["收藏"],
        "summary": "更新收藏备注",
        "description": "更新收藏商品的备注信息",
        "parameters": [
          {
            "name": "favorite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/search-records": {
      "get": {
        "tags": ["搜索记录"],
        "summary": "获取搜索记录",
        "description": "获取用户的搜索历史记录，按时间倒序排列",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "页码，从1开始",
            "example": 1
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "每页数量，最大100条",
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}